feat: implement inelegant solution for footnotes
Problem: At the time of writing, Prismic documents don’t support in-page
links, which means we can’t have footnotes that are linked-to from
within a weeknote.

Solution: It’s not very elegant but by parsing the Prismic-generated
weeknote HTML, we can manipulate it into having linkable footnotes. Within
a Prismic document, writing `[n]` (any number) anywhere in the document
will be transformed into a link to the corresponding list item index of
footnotes, if a list is found directly adjacent to a
`<h2>Footnotes</h2>` heading. Links back to the sources are also added. Meh, it works.