fix: strip footnote references from weeknote lists
It doesn’t really make sense to show something like this is:

- weeknote C
- weeknote B[1]
- weeknote A

The “[1]” is a reference to a footnote within the individual
weeknote, but that only really makes sense when you’re reading that
weeknote. In this list format, it seems to make sense to remove it.

Note that I’m keeping the footnote references intact for the RSS version of
the weeknotes list, because RSS readers tend to use the same weeknote title
in two different contexts: an inbox feed view (as a list of posts), and
as the main title of an individual post. So in order to keep the
footnote references in the individual weeknote’s main heading of an RSS
weeknote, we need to also keep it for the RSS list view and just accept that as a pooey
constraint of RSS. (Maybe I’m just missing a trick though 🤷!)